+++ lottery-core/src/test/java/de/alpharogroup/lottery/combinations/CombinationResolverTest.java
@@ -24,7 +24,12 @@ import static org.testng.AssertJUnit.assertEquals;
 import static org.testng.AssertJUnit.assertTrue;
 
 import java.util.List;
+import java.util.Set;
 
+import de.alpharogroup.collections.set.SetFactory;
+import de.alpharogroup.lottery.LotteryExtensions;
+import de.alpharogroup.lottery.box.DoubleCollectionBox;
+import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
 import de.alpharogroup.collections.CollectionExtensions;
@@ -35,6 +40,102 @@ import de.alpharogroup.collections.list.ListFactory;
  */
 public class CombinationResolverTest
 {
+
+	/**
+	 * Test the method {@link CombinationResolver#getPossibleCombinations(List, int, List, int)}
+	 */
+	@Test
+	public void testGetPossibleCombinations() {
+		List<Integer> firstCollection;
+		int firstCollectionCombinationSize;
+		List<Integer> secondCollection;
+		int secondCollectionCombinationSize;
+		Set<DoubleCollectionBox<Integer>> actual;
+		Set<DoubleCollectionBox<Integer>> expected;
+			// new scenario...
+		firstCollection = ListFactory.newArrayList(1, 2, 3, 4, 5, 6);
+		firstCollectionCombinationSize = 5;
+		secondCollection = ListFactory.newArrayList(1, 2, 3);
+		secondCollectionCombinationSize = 2;
+		actual = CombinationResolver.getPossibleCombinations(firstCollection, firstCollectionCombinationSize,
+			secondCollection, secondCollectionCombinationSize);
+		expected = SetFactory.newHashSet();
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 2, 4, 5, 6))
+			.secondCollection(SetFactory.newHashSet(1, 2)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 2, 4, 5, 6))
+			.secondCollection(SetFactory.newHashSet(1, 3)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 2, 4, 5, 6))
+			.secondCollection(SetFactory.newHashSet(2, 3)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 2, 3, 5, 6))
+			.secondCollection(SetFactory.newHashSet(1, 2)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 2, 3, 5, 6))
+			.secondCollection(SetFactory.newHashSet(1, 3)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 2, 3, 5, 6))
+			.secondCollection(SetFactory.newHashSet(2, 3)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 2, 3, 4, 6))
+			.secondCollection(SetFactory.newHashSet(1, 2)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 2, 3, 4, 6))
+			.secondCollection(SetFactory.newHashSet(1, 3)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 2, 3, 4, 6))
+			.secondCollection(SetFactory.newHashSet(2, 3)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 2, 3, 4, 5))
+			.secondCollection(SetFactory.newHashSet(1, 2)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 2, 3, 4, 5))
+			.secondCollection(SetFactory.newHashSet(1, 3)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 2, 3, 4, 5))
+			.secondCollection(SetFactory.newHashSet(2, 3)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(2, 3, 4, 5, 6))
+			.secondCollection(SetFactory.newHashSet(1, 2)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(2, 3, 4, 5, 6))
+			.secondCollection(SetFactory.newHashSet(1, 3)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(2, 3, 4, 5, 6))
+			.secondCollection(SetFactory.newHashSet(2, 3)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 3, 4, 5, 6))
+			.secondCollection(SetFactory.newHashSet(1, 2)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 3, 4, 5, 6))
+			.secondCollection(SetFactory.newHashSet(1, 3)).build());
+		expected.add(DoubleCollectionBox
+			.<Integer>builder()
+			.firstCollection(SetFactory.newHashSet(1, 3, 4, 5, 6))
+			.secondCollection(SetFactory.newHashSet(2, 3)).build());
+		assertEquals(18, actual.size());
+		assertEquals(expected.size(), actual.size());
+		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
+	}
 
 	/**
 	 * Test the method {@link CombinationResolver#getAllCombinations(List, int)}
@@ -53,6 +154,8 @@ public class CombinationResolverTest
 		expected.add(ListFactory.newArrayList(1, 2, 4));
 		expected.add(ListFactory.newArrayList(1, 3, 4));
 		expected.add(ListFactory.newArrayList(2, 3, 4));
+		assertEquals(4, actual.size());
+		assertEquals(expected.size(), actual.size());
 		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
 		// new scenario...
 		expected = ListFactory.newArrayList();
@@ -68,6 +171,8 @@ public class CombinationResolverTest
 		expected.add(ListFactory.newArrayList(2, 3, 5));
 		expected.add(ListFactory.newArrayList(2, 4, 5));
 		expected.add(ListFactory.newArrayList(3, 4, 5));
+		assertEquals(10, actual.size());
+		assertEquals(expected.size(), actual.size());
 		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
 		// new scenario...
 		expected = ListFactory.newArrayList();
@@ -80,6 +185,8 @@ public class CombinationResolverTest
 		expected.add(ListFactory.newArrayList(1, 2, 4, 5, 6, 7));
 		expected.add(ListFactory.newArrayList(1, 3, 4, 5, 6, 7));
 		expected.add(ListFactory.newArrayList(2, 3, 4, 5, 6, 7));
+		assertEquals(7, actual.size());
+		assertEquals(expected.size(), actual.size());
 		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
 		// new scenario...
 		values = ListFactory.newRangeList(1, 11);
@@ -91,7 +198,7 @@ public class CombinationResolverTest
 	 * Test the method {@link CombinationResolver#getCombinations(List, int)}
 	 */
 	@Test
-	public void testGetCombinations()
+	public void testGetCombinationsWithStringList()
 	{
 		List<String> values;
 		List<List<String>> actual;
@@ -104,6 +211,8 @@ public class CombinationResolverTest
 		expected.add(ListFactory.newArrayList("1", "2", "4"));
 		expected.add(ListFactory.newArrayList("1", "3", "4"));
 		expected.add(ListFactory.newArrayList("2", "3", "4"));
+		assertEquals(4, actual.size());
+		assertEquals(expected.size(), actual.size());
 		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
 		// new scenario...
 		expected = ListFactory.newArrayList();
@@ -119,6 +228,8 @@ public class CombinationResolverTest
 		expected.add(ListFactory.newArrayList("2", "3", "5"));
 		expected.add(ListFactory.newArrayList("2", "4", "5"));
 		expected.add(ListFactory.newArrayList("3", "4", "5"));
+		assertEquals(10, actual.size());
+		assertEquals(expected.size(), actual.size());
 		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
 		// new scenario...
 		expected = ListFactory.newArrayList();
@@ -131,11 +242,93 @@ public class CombinationResolverTest
 		expected.add(ListFactory.newArrayList("1", "2", "4", "5", "6", "7"));
 		expected.add(ListFactory.newArrayList("1", "3", "4", "5", "6", "7"));
 		expected.add(ListFactory.newArrayList("2", "3", "4", "5", "6", "7"));
+		assertEquals(7, actual.size());
+		assertEquals(expected.size(), actual.size());
 		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
 		// new scenario...
 		values = ListFactory.newArrayList("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11");
 		actual = CombinationResolver.getCombinations(values, 3);
 		assertEquals(actual.size(), 165);
+	}
+
+	/**
+	 * Test the method {@link CombinationResolver#getCombinations(List, int)}
+	 */
+	@Test
+	public void testGetCombinationsWithIntegerList()
+	{
+		List<Integer> values;
+		List<List<Integer>> actual;
+		List<List<Integer>> expected;
+		// new scenario...
+		expected = ListFactory.newArrayList();
+		values = ListFactory.newArrayList(1, 2, 3, 4);
+		actual = CombinationResolver.getCombinations(values, 3);
+		expected.add(ListFactory.newArrayList(1, 2, 3));
+		expected.add(ListFactory.newArrayList(1, 2, 4));
+		expected.add(ListFactory.newArrayList(1, 3, 4));
+		expected.add(ListFactory.newArrayList(2, 3, 4));
+		assertEquals(4, actual.size());
+		assertEquals(expected.size(), actual.size());
+		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
+		// new scenario...
+		expected = ListFactory.newArrayList();
+		values = ListFactory.newArrayList(1, 2, 3, 4, 5);
+		actual = CombinationResolver.getCombinations(values, 3);
+		expected.add(ListFactory.newArrayList(1, 2, 3));
+		expected.add(ListFactory.newArrayList(1, 2, 4));
+		expected.add(ListFactory.newArrayList(1, 2, 5));
+		expected.add(ListFactory.newArrayList(1, 3, 4));
+		expected.add(ListFactory.newArrayList(1, 3, 5));
+		expected.add(ListFactory.newArrayList(1, 4, 5));
+		expected.add(ListFactory.newArrayList(2, 3, 4));
+		expected.add(ListFactory.newArrayList(2, 3, 5));
+		expected.add(ListFactory.newArrayList(2, 4, 5));
+		expected.add(ListFactory.newArrayList(3, 4, 5));
+		assertEquals(10, actual.size());
+		assertEquals(expected.size(), actual.size());
+		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
+		// new scenario...
+		expected = ListFactory.newArrayList();
+		values = ListFactory.newArrayList(1, 2, 3, 4, 5, 6, 7);
+		actual = CombinationResolver.getCombinations(values, 6);
+		expected.add(ListFactory.newArrayList(1, 2, 3, 4, 5, 6));
+		expected.add(ListFactory.newArrayList(1, 2, 3, 4, 5, 7));
+		expected.add(ListFactory.newArrayList(1, 2, 3, 4, 6, 7));
+		expected.add(ListFactory.newArrayList(1, 2, 3, 5, 6, 7));
+		expected.add(ListFactory.newArrayList(1, 2, 4, 5, 6, 7));
+		expected.add(ListFactory.newArrayList(1, 3, 4, 5, 6, 7));
+		expected.add(ListFactory.newArrayList(2, 3, 4, 5, 6, 7));
+		assertEquals(7, actual.size());
+		assertEquals(expected.size(), actual.size());
+		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
+		// new scenario...
+		values = ListFactory.newRangeList(1, 11);
+		actual = CombinationResolver.getCombinations(values, 3);
+		assertEquals(actual.size(), 165);
+		// new scenario...
+		values = ListFactory.newRangeList(1, 8);
+		actual = CombinationResolver.getCombinations(values, 6);
+		assertEquals(actual.size(), 28);
+	}
+
+	/**
+	 * Test the method {@link CombinationResolver#getAllPossibleCombinationsCount(int, int, int, int)}
+	 */
+	@Test
+	public void testGetAllPossibleCombinationsCountWithFourParams(){
+
+		List<Integer> firstValues;
+		List<Integer> secondValues;
+		long actual;
+		long expected;
+		// new scenario...
+		firstValues = ListFactory.newArrayList(1, 2, 3, 4, 5);
+		secondValues = ListFactory.newArrayList(1, 2, 3);
+		actual = CombinationResolver.getAllPossibleCombinationsCount(firstValues.size(), 3,
+			secondValues.size(), 2);
+		expected = 30;
+		assertEquals(actual, expected);
 	}
 
 	/**
@@ -196,6 +389,16 @@ public class CombinationResolverTest
 			* allPossibleCombinationsEurojackpotStarNumbersCount;
 		expected = 95344200;
 		assertEquals(actual, expected);
+	}
+
+	/**
+	 * Test method for {@link CombinationResolver} with {@link BeanTester}
+	 */
+	@Test
+	public void testWithBeanTester()
+	{
+		final BeanTester beanTester = new BeanTester();
+		beanTester.testBean(CombinationResolver.class);
 	}
 
 }